6239-fix-remote-ip-logging-and-trusted-proxy-config (#4576)
* 6239-rate-limiting-prep-config

* 6239-rate-limiting-prep

just use rails logger

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep

* 6239-rate-limiting-prep
