+++ config/application.rb
@@ -12,6 +12,7 @@ ENV['RAILS_DISABLE_DEPRECATED_TO_S_CONVERSION'] = 'true'
 Bundler.require(*Rails.groups)
 
 require_relative '../lib/util/id_protector'
+require_relative '../lib/util/rails_trusted_proxies_config'
 
 module BostonHmis
   class Application < Rails::Application
@@ -22,6 +23,12 @@ module BostonHmis
     # ActionCable
     config.action_cable.mount_path = '/cable'
     config.action_cable.url = ENV.fetch('ACTION_CABLE_URL') { "wss://#{ENV['FQDN']}/cable" }
+
+    ENV['TRUSTED_PROXIES'].presence&.then do |trusted_proxies|
+      parsed = RailsTrustedProxiesConfig.parse_csv(trusted_proxies)
+      # if we are adding custom trusted proxies, we need to include the default addrs
+      config.action_dispatch.trusted_proxies = (ActionDispatch::RemoteIp::TRUSTED_PROXIES + parsed) if parsed
+    end
 
     Rails.application.config.active_record.belongs_to_required_by_default = true
     # https://discuss.rubyonrails.org/t/cve-2022-32224-possible-rce-escalation-bug-with-serialized-columns-in-active-record/81017
